대시보드 응답 타입과 null 정책을 통일하고 거래 데이터 제약 강화
- 금액 필드를 BigDecimal 기반 숫자로 통일하고 시세 누락 시 부분 응답 대신 오류 반환
- 소유주 ID를 Long으로 변경하고 증권사 필수/수량 Int 범위를 Flyway 마이그레이션에 반영
- 프론트 스키마와 렌더링, OpenAPI 및 회귀 테스트를 변경된 구조에 맞게 수정
- Kotlin not null 타입과 Jackson Kotlin 모듈로 OpenAPI 필수 필드를 자동 추론하고 중복 Swagger 어노테이션 정리

+++ backend/src/main/kotlin/kr/byeongmin/stockdaejang/domain/owner/entity/Owner.kt
@@ -13,7 +13,7 @@ class Owner(
     @Id
     @Column(name = "id", nullable = false)
     @field:Schema(description = "소유주 내부 대리키", example = "1", minimum = "1")
-    val id: Short,
+    val id: Long,
 
     @Column(name = "name", nullable = false, unique = true)
     @field:Schema(description = "소유주명", example = "병민")
