+++ backend/src/main/kotlin/kr/byeongmin/stockdaejang/domain/trade/dto/ParsedPositionDto.kt
@@ -1,7 +1,7 @@
 package kr.byeongmin.stockdaejang.domain.trade.dto
 
 internal data class ParsedPositionDto(
-    val ownerId: Short,
+    val ownerId: Long,
     val brokerageCode: String,
     val itemCode: String,
 )
